pyproject.toml
[tool.poetry]
name = "django-graphql-federation-template"
version = "1.0.0"
description = "Django + FastAPI + Strawberry GraphQL Federation Template"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.0"
strawberry-graphql = "^0.252.0"
pyjwt = "^2.8.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gunicorn = "^22.0.0"
requests = "^2.32.0"
fluent-logger = "^0.11.0"
pytest = "^8.1.0"
syrupy = "^4.6.0"
httpx = "^0.27.0"
boto3 = "^1.34.0"
pydantic = "~2.6.4"
pydantic-settings = "^2.2.0"
ua-parser = "^0.18.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"
# Add your custom dependencies here
python-dotenv = "^1.0.1"
prometheus-fastapi-instrumentator = "^7.0.0"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
graphql-relay = "^3.2.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.1.0"
black = "^24.4.0"
safety = "^3.1.0"
isort = "^5.13.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
